fancontrold: add hysteresis for fan fault condition
Do not try to power off the system if we get only one alarm of a faulty
fan. This patch adds a new command line parameter to set a counter for
the number of faults we need to see in a row, before powering off the
system. The default value is 5.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>